Fix timestamp used when promoting replica to main (#2366)
When replica gets promoted to the main, its timestamp
(storage->timestamp_) remains the same or takes the timestamp value that
was used for committing data received from the previous main (last
durable timestamp). The issue was that timestamp semantically means next
available timestamp and the last durable timestamp wasn't incremented.