Bugfix: Forcing streams and tll to stop before destroying InterpreterContext (#2361)
Streams and TTL have their own interpreters that run in the background.
These interpreters prepare queries and execute them.
When shutting down, the InterpreterContext would be destroyed before
stopping streams and TTL.
There was a possibility they would be preparing queries while the IC was
being destroyed.
While Pulls are protected by the abort signal, Prepare is not.