Handle out of order and stale RPC messages in high availability (#2373)
Added logs on several important methods to improve observability and
easen debugging. Previously, when registering a replica and when one of
databases wasn't registered successfully, the registration would still
pass although this would create problems in the system. Now the
registration will fail if that happens. Registering replica in HA could
fail because the instance is not main anymore. Instead of crashing the
database, error code NOT_MAIN will now be returned in order to cope with
such requests. On failed registration (any reason), we always tried to
unregister replica. However, we never checked the status of such
unregistration ( = undo operation). Now, if unregistration fails because
of disk issue, we will crash the database since the system in
unrecoverable and if unregistration passes, the error code from
registering replica is returned. GetReplicaUUID won't crash DB if it
received request as MAIN, it will rather just return empty optional. In
the test code, we try to fetch data needed for choosing instance to kill
from all 3 coordinators, not just from n4 as was previously arbitrarily
done.