Bugfix: Use one type of Raft logs and remove distributed locks. (#2430)
Coordinators now use a single type of raft logs (update_cluster_state)
to remove the need for distributed locks. Coordinators can now be in 3
states: follower, leader not ready to react to instance callbacks and
leader which is ready to react to instance callbacks. When a coordinator
becomes the leader, it sets its state to leader not ready and then
changes it to ready only after reconciliation finishes.