Planner to replace point.distance filter with index scan (#2362)
Make planner utilize spatial index when querying with point.distance
function.

Example:
`WITH point({x:0,y:0}) AS a MATCH (m:L1) WHERE point.distance(m.prop, a)
< 1 RETURN m;`

Currently the index is only utilized when the point its querying around
(`a` in the query above) is an identifier (is extracted with WITH clause
before). In one of the next PRs the planner will be changed to
automatically extract constants to identifiers.

---------

Co-authored-by: imilinovic <ivan.milinovic123@gmail.com>