Bugfix: Deadlock when demoting data instance (#2447)
When a coordinator sends a request for promotion or demotion, the lock
on replication state is taken so that it cannot be interleaved with
commit. If main became a replica and was executing write txn, the txn
will be aborted. Scheduler is now stopped before thread pool in rpc
client.