Handle failure in WAL replication (#2467)
Time for running Jepsen stress test increased to 3h. When replica
receives a WAL file and has to create a vertex, it won't crash database
if such vertex already exists. Rather it will throw an exception which
will abort replication transaction and return the information to main
instance which sent the WAL request. Main instance will then decide
whether replica needs to go to recovery or some other state. If main
becomes a replica with a write txn to commit, instead of calling Abort
directly, the code will throw when exception and the interpreter will
then abort through that exception. Otherwise we have a problem with
transaction status.