Align recovery timestamp check with timestamp modifications (#2476)
We are checking the last durable timestamp the end of the recovery
process.
The check determines if the recovery was successful. 
Since the recovery process is async, we could be recovering while
committing and updating the timestamp.
Now we lock the engine_lock_ when checking the timestamp and make sure
the view is consistent.