Remove EdgeSetFrom, EdgeSetTo, EdgeChangeType (#2530)
These have been deemed low value and add extra complexity to other parts
of the code base.
Its an important invariant that an edge during its lifetime is only ever
be owned by same vertices as when the edge was created.
This is important for the correct handling of edges. Among other things,
indexing and edge deletion require that the edge's type, source and
destination are invariant.

---------

Co-authored-by: Gareth Lloyd <gareth.lloyd@memgraph.io>