Fix client-side routing in K8s HA deployment scenario (#2568)
Up to this point, the 1st leader couldn't find out information about its
bolt server. In order to support that, users will run `ADD COORDINATOR
1..` additionally to `ADD COORDINATOR 2..` and `ADD COORDINATOR 3...`.
Informations about coordinators in NuRaft cluster are saved at two
places. Information about coordinator server and management server are
saved inside `raft_server::aux` while information about bolt server
inside application logs. The process of adding a coordinator (not self)
then consists of two steps. In the 1st step we add coordinator to the
networking part of NuRaft. In the 2nd step, we create an application log
which stores information about the current state of the cluster. If the
1st step passed but the 2nd step didn't pass, we crash the database
since this is an invalid situations. This situation should never occur
at the startup but it serves us as a protection mechanism to remain
consistent and safe. Data structures `CoordinatorInstanceAux` and
`CoordinatorInstanceContext` are used to represent coordinator state in
application logs and as part of NuRaft networking configuration.

Files from nuraft/ and coordination/ are all moved to coordination/
since there is no more clear boundary between files in each folder.

Flag `--instance-uuid-frequency-sec` is removed since it's not used
anymore (it should have been deprecated past release). Flags
`--instance-down-timeout-sec` and
`--instance-health-check-frequency-sec` are still used but moved under
ownership of coordinator instance.

When run on empty cluster (cluster on which neither of `ADD COORDINATOR`
queries were run), queries `SHOW INSTANCES` and `SHOW INSTANCE` will
return the coordinator on which that query is run. Correct bolt server
will be returned if `ADD COORDINATOR` is run on that coordinator and if
not, empty string will be returned.

SSL is removed from all configurations since we don't support it yet. 

Coordinator IDs are represent using int32_t since this is used by
NuRaft. Until now we were using uint32_t and doing unnecessary casts.