Fix edge type index cleanup on Abort (#2596)
Abort does cleanup of objects and indexes. To work out what entries to
remove from the index the current value of the entry needs to be figured
out from the object as it is and the delta. The edge type information is
held on the Vertex, not the Edge, hence need to process all edges first,
otherwise we loose the information held on vertex which is needed to
know which index entries need cleaning up.