Change query modules allocator (#2561)
- Change query modules to use multipool instead of monotonic allocator
because monotonic would sometimes cause OOM on query modules that create
and delete a lot of objects due to not being able to reclaim memory.
- Add reference type to C++ API and change MGP::List and MGP::Map get
operations to return a reference.
- Removed noexcept from C++ API calls that didn't make sense.

BREAKING: Incorrect usage of mgp_value (C - API) now causes a crash -
you must be very careful not to do a double deallocation. `mgp_value`
has a destructor and therefore `mgp_value_destroy` must only be called
on a pointer. Also after running `mgp_value_destroy` you MUST not call
mgp_`type`_destroy on the underlying data because that will cause double
deallocation and a crash.

---------

Co-authored-by: Gareth Lloyd <gareth.lloyd@memgraph.io>