Improving replication recovery process (#2617)
The semantic of snapshot handler changed in the following way: Either
handling snapshot request passes or it doesn't. If it passes we return
the current commit timestamp of the replica. If it doesn't pass, we
return optional which will signal to the caller that it shouldn't update
the commit timestamp value. If loading snapshot file failed, it won't
crash the database anymore. If an error happened during snapshot
recovery, instead of crashing the database, we will now log a message
and clear the storage. For recovery using WALFile, commit timestamp on
main's side shouldn't be updated if:
1.) the database accessor couldn't be obtained
2.) UUID sent with the request is not the current MAIN's UUID which
replica is listening to
Only if loading all WAL files succeeded then main can continue recovery
if it should send more recovery steps. If loading WAL files partially
succeeded then we cannot continue recovery because CurrentWAL could come
at the wrong place in the replication queue. Similarly, recovery using
CurrentWAL is only considered successful if it passed completely.
Loading WAL file won't crash the database if the WAL file cannot be
loaded. The recovery task won't spin in a while loop anymore until it
completely recovers the replica. Instead, it will generate recovery
steps, try to recover replica and set the state of replica to READY or
MAYBE_BEHIND depending on the progress achieved. The state will be
MAYBE_BEHIND if the MAIN committed since the start of the recovery.

Before this PR, the ASYNC replication could deadlock since the same
queue is used for both recovery and commit tasks. This can be a problem
since both recovery and commit use the same rpc lock in the situation
when commit task is added after the recovery task. In recovery task, we
wait for commit task to release RPC lock but commit task is never
executed because it is behind a single-threaded queue. This is now fixed
by recovery task not executing if the replica has already been
recovered.