Protect the iterator lifetime with shared_ptr (#2633)
The recent changes in the memory allocator used for user defined
functions (#2561), has revealed issues in our C++ API. There were double
deallocations of objects that with the previous monotonic allocator were
not an issue (the memory still existed). This fix extends the lifetime
of those objects to avoid crashes with the new allocator.

The C++ API should be overhauled at some point soon.