Fix subtle data race error in SkipList
find_equal_or_greater_ can be used to find first entry that matches a
prefix/value. If it was marked delete at the wrong time, nullptr would
be returned, which logically means empty iteration. But it is possible
that the next value in skip_list would be fine.
Fix is to retry the find_node.