Fix data race in skip list usage (#2686)
Issue was initially found with unique constraint. But root cause was
with our `find_equal_or_greater`. The data race is most likely only just
become noticeable because of recent performance changes that as a result
increased the likelihood this data race could be observed.

Also delivered:
- SkipList pretty printer for gdb
- unit test to validate find_equal_or_greater behaviour
- fix for data race
- bugfix for when unique constraints validation for vertex with multiple
labels
- optimization to avoid work in commit when no existence constraints 
- optimization to avoid work in commit when no unique constraints 
- simplify cleanup after unique constraint violation
- optimization extraction of multiple properties from property store
done in a single pass

---------

Co-authored-by: Colin Barry <colin.barry@memgraph.io>