Return newest index when creating nodes in jepsen HA test (#2678)
When creating nodes, returns the newly created index to enable
debugging. Since the query for creating nodes doesn't expose variables
used, we need a way to connect batches with time. With this change we
will be able to see in jepsen.log at which point in time some batch was
created.