Add dynamic relationship creation in Cypher (#2558)
The edge type can now be a symbolic name, parameter or property lookup.
The
implication of this is that it's easier to create more edge types under
a
single graph.

Use-cases:
* import of many edge types, with this feature, you need only a single
query
    to create many edge types
* if you encode info as edge types, not as properties (if you only want
a
    small number of distinct CREATE queries), your query performance can
    increase (in general) because the processing of edge types is faster
compared to dealing with properties, but that might not always be the
case
AKA you need a deep understanding of your workload to get the
performance
    benefits.