Bugfix: Ignore hidden files when symlinking data directory (#2748)
Memgraph can be started in K8s environment with various storage
providers. Some of them create `.snapshot` file which was preventing the
database from starting. When Memgraph is started in a single-tenant
mode, durability files are linked to those from the root for the
back-compatibility. All hidden files should be ignored.