Bugfix: force reset causes data loss and deletion of durability files in HA environment (#2727)
Running time of the Jepsen stress test reduced to 7h because of large
logs. ForceResetRPC is removed from the cluster and its functionality is
merged with the functionality of the recovery process. If the branching
point is found, with the the 1st recovery step, the additional
information is sent which the replica will understand as instruction to
clear its storage. At that point, all old durability files are moved
inside .old directory. This means that replica doesn't respect
.snapshot_retention_count configuration parameter. Files are preserved
inside the .old directory so that user can have all its files there in
the case the forked timeline event occurs. The replication state
DIVERGED_FROM_MAIN now only exists for the community and enterprise
replication, in the HA, the replica will go immediately to the RECOVERY
state.