Bugfix: Storage access timeout allowed everywhere (#2779)
Recent changes made possible for Storage::Access to timeout.
This is needed for Cypher queries, as their execution is split into
Prepare and multiple Pulls. Not timing out could lead to a deadlock.
On the other hand, this is not fine in other circumstances (ex.
SnapshotCreate).
This fixes the issue by allowing the access timeout only for Prepare.