Bugfix: Avoid throwing in slk when a buffer isn't read fully (#2770)
slk::Reader would throw before if data from the buffer wouldn't be
completely ready. This caused network connections between main and
replica to be constantly dying when unnecessary. Now when replica
doesn't read all data from the buffer, the exception thrown in the slk
part of the code is handled properly, avoiding the connection drop.
Additionally, the PR tries to reduce the massive amount of logs produced
in HA code when trace log level is used. The PR also improves jepsen's
run script so that it doesn't fail when jepsen.log cannot be read. With
this fix, we should be able to see at least part of the logs which can
also be helpful for debugging. Jepsen tests are also checking whether
`There is still leftover data in the SLK stream` exception will be
thrown during the execution.