Bugfix: Prevent deadlock on data instance (#2816)
The following set of events caused a deadlock in our HA:

1. MAIN receives query to commit as implicit txn. Main lock is hold all
the time because of query,
2. While executing and before the commit, main is demoted to replica.
3. Replica receives snapshot rpc and tries to acquire the main lock.
4. Replica receives request to get promoted. RPC server cannot be closed
because snapshot rpc isn't finished.

The fix consists of waiting for a lock in replication handlers for
SnapshotRpc, WalFilesRpc and CurrentWalRpc but with a predefined
timeout. Additionally, log message for addrinfo is now added, same as
log message when trying to close Socket.