test: Fix query_dump tests for on-disk
In-memory indices optimise the case of "impossible" ranges where the
lower and upper bounds are of different types. For example, the
following query can never yield any vertices:

```cypher
MATCH (n:L1) WHERE n.a > 42 AND n.a < "hello" RETURN *;
```

This case is handled by reporting the invalid range, a state detected by the
in-memory indices. But because the on-disk indices don't handle this special
case (yet), the bounds for an invalid range are now passed through to the
in-memory/on-disk index.