fix: Check lower bounds only after iterator moved by find_equal_or_greater
Because `find_equal_or_greater` ignores the inclusive or exclusive
nature of the lower bound, the iterator cannot skip the lower bound
check for the very first call to `AdvanceUntilValid`.

Our previous optimisation introduced the following error:
```cypher
CREATE INDEX ON :L1(a);
CREATE (:L1 {a: 0}), (:L1 {a: 1}), (:L1 {a: 2});
MATCH (n:L1) WHERE n.a > 0 AND n.a <= 2 RETURN n;

+--------------+
| n            |
+--------------+
| (:L1 {a: 0}) |
| (:L1 {a: 1}) |
| (:L1 {a: 2}) |
+--------------+
```

The optimisation is still valid in the majority of cases, but for the
very first `AdvanceUntilValid`, the lower bound will also be checked so
that nodes matching exclusive lower bounds are discarded.