refactor: Remove `vertex_cache_count` dependency on `TypedValue`
`TypedValue` was used here because `PropertyValue`s needed to be hashed
in order to use them as keys in the cache map. As `PropertyValue` didn't
then support hashing, instead they were converted to `TypedValue` (which
did). No longer need to do this conversion as `PropertyValue`'s can now
be hashed.