test: Remove over-specific check for types using < operator
Previously, due to the way that the
ScanAllByLabelPropertiesRangeOperator worked, it would throw an
exception if either of the operands to `WHERE node.x < y` was an
"incomparable" type.

This is no longer an issue - the new scan operator handles it by
returning an empty record set - and so the tests are no longer
relevant.