fix: Treat incomparable ranges the same in disk and inmemory
Our in-memory iterators can handle the case of mixed lower and upper
bounds, whilst the disk iterators cannot. This adds a guard to in-memory
implementation to unify the behaviour for both (a less risky change than
updating the disk iterators!), and puts back a test which was previously
failing due to the divergent behaviour.