Eagerly return if registering one of replicas fail upon promotion. (#2869)
When a replica gets promoted to main, it tries to register all other
replicas. If registering one of replicas failed, there was a bug in SLK
where the buffer would contain both false + true. When an error is
encountered during promotion, promote is aborted and this information is
returned back to coordinator. The coordinator should then try to repeat
the registration process.