Bugfix: New leader doesn't elect a main in the case when there are only replicas in Raft logs (#2857)
A coordinator can become leader with the raft state containing only
replica instances. This is possible if user demoted main instance and
then coordinator crashed. In that case, the new leader should do a
failover but this wasn't before.
When promoting instance, data instances can now be down and promote will
still succeed. Previously, all data instances had to be alive for manual
promotion to be done.