Bugfix: Scheduler cannot be paused while executing its function (#2882)
The PR fixes the following situation:
There are 2 threads, 1st one is the periodic snapshot thread running
from the scheduler, the 2nd one is the demotion thread running in
replication_handler.

T1 takes unique_lock in scheduler and holds it during the function run.
T2 takes unique repl state lock and tries to Pause the periodic snapshot
thread. Pausing the snapshot thread requires unique lock in the
scheduler, held by T1. T1 in the meantime tries to take repl state lock
to read the replication role so it could decide whether to take the
snapshot or not (depending on whether it is main or replica).

The correct thing is to release the lock in scheduler before starting
snapshot creation.