Fix flaky Kafka consumer unit tests (#2881)
This PR removes a fragile and flaky assertion in the `BatchInterval`
test that was checking whether each batch was processed strictly within
a `batch_interval` ± 10% window. In practice, we can't really know or be
certain that a message will be processed exactly on each interval, there
are delays from the Kafka mock (e.g., librdkafka), thread wake-ups, OS
scheduling, etc. Even if we increase the `batch_interval`, we still
don’t know where the actual boundary is or what interval would make this
stable. And when it fails, we don’t get any useful signal about what
went wrong or if something even is wrong. Because of that, the check
that all expected messages were received (with correct content) is much
more important and that’s already covered in the test. So the flaky
timing check was removed to make the test more reliable while still
validating the important behavior.