Label+Properties Composite Index (#2760)
Add functionality for composite index on vertices.

Extending label + property index to have multiple properties. `CREATE
INDEX ON :L(a,b,c);`. The ordering of the properties matter, `:L(a,b)`
is different from `:L(b,a)`.

The index is used in planning to reduce the cost of scan + filter, by
using the lowest costing index to improve performance of the scan.

breaking: SHOW INDEX INFO; will now have a list for label+properties
index

---------

Co-authored-by: Colin Barry <colin.barry@memgraph.io>