Improve the failover process in HA when using multitenancy (#2812)
The failover in the MT code now has a different logic. We 1st try to
find the instance which has most newest databases. If there are multiple
instances that are newest on the same number of databases, the sum of
timestamps of all databases decides. The instance which has larger sum
it is assumed to be newer.