Fix release build tests (#2900)
Make daily builds run the large stress test and release Jepsen test.
- Jepsen was being skipped because the input OS was being set to
`ubuntu-24.04` - this input is now ignored
- Large stress test was being skipped because `run_stress_large` was not
set in the `TestBuild` job, this is now fixed
- removed `cron` job for stress test large, as this will take it's place

The downside of this is that daily builds will now take 8-9 hours before
triggering MAGE builds; the upside is that the previously skipped tests
will be included in the aggregated test result and therefore count
towards the overall result displayed on the daily build page.