User-configurable option to failover only to SYNC replica (#2931)
Using the query `set coordinator setting 'sync_failover_only' to
'true'/'false'` users can modify the failover logic when deciding on
which instance to perform a failover. When the option is set to true,
only sync replicas will be considered. The option is by default set to
true to minimize the chances of experiencing a data loss upon failover.