Add timeout on socket connect (#2949)
The PR adds the functionality of socket connect with timeout. The only
portable way of doing this is creating a socket in a non-blocking mode,
timeout check using `poll` and switching it back to a blocking mode
since we all our sockets are used in blocking mode. UNIX doesn't have an
API which would allow that from the start. The primary motivation for
such a change is the fact that socket connect timeout can be quite high
(~130s) and that it depends on the platform and on the number of SYN
packets sent during connect. Two properties need to be satisfied for a
system to be considered synchronous:
1. There is a known upper bound on processing delays. This includes the
time needed to receive a message, perform a local computation and
replying
2. There is a known upper bound on message transmission delays.

We already had timeouts on read and write message, the only thing left
was a timeout on connect.

The timeout for StateCheckRpc and FrequentHeartbeatRpc is reduced to 5s.