Bugfix: Fix replication state machine when recovery and commit tasks are interleaved (#2973)
The PR fixes the bug with asynchronous replication. If the recovery
would've finished between starting and finishing txn, it would set the
replica state to ready and then a txn wouldn't ever be replicated. This
is fixed by protecting the exit from FinalizeTransactionReplication
using engine lock.

The replication performance is improved by reducing the memory order
level required from sequentially consistent to acquire.