Add bolt+routing connection to HA create Jepsen test (#2835)
The test `ha_create` now uses arguments `--nemesis-start-sleep` and
`--recovery-time`. All queries now go through the current leader which
returns a routing table to the client. The client then propagates write
queries to the current main in the cluster. There was a possible case in
which the current main would execute in parallel multiple requests which
could've led to data duplication but this is now handled using unique
constraint to ensure serializability of the data.