Bugfix: Fix deadlock when using ASYNC replicas by skipping recovery tasks if lock cannot be obtained (#2984)
When using ASYNC replicas, MAIN could deadlock because in its thread
pool there would be a commit task holding RPC lock. Commit task would
wait for other tasks to finish but these tasks wouldn't not be able to
get executed because RPC lock couldn't be acquired. The current fix is
to try lock on RPC lock and if this is not possible, skip the task. This
shouldn't be a problem because replica will stay in RECOVERY state,