Remove SnapshotDigest from InMemoryAnalytical (#3000)
In [PR2910](https://github.com/memgraph/memgraph/pull/2910)
SnapshotDigest was introduced to not allow creating a snapshot if its
the same as previous one. It uses `last_durable_timestamp` which is only
valid in IN_MEMORY_TRANSACTIONAL which led to snapshots not being able
to get triggered in IN_MEMORY_ANALYTICAL due to SnapshotDigest always
being the same.

No changelog as this wasn't released.