Support IN clauses in label+property index queries (#3014)
Fix for Label+property queries using `IN` clauses returning no results.

If we use an `IN` query without a label+property index, the query
returns the result as expected:

```cypher
CREATE (:L1 {a: 42});
MATCH (x:L1) WHERE x.a IN [42] RETURN x.a;
```
```
+-----+
| x.a |
+-----+
| 42  |
+-----+
```

But if the query planner uses an index, the `IN` no longer works and an
empty result set is instead retrieved:
```cypher
CREATE INDEX ON :L1(a);
MATCH (x:L1) WHERE x.a IN [42] RETURN x.a;
```
```
Empty set
```

We expect the same results to be returned if we have a usable index for
the query.

---------

Co-authored-by: Gareth Lloyd <gareth.lloyd@memgraph.io>