Enabling replication in case of RECOVER SNAPSHOT (#3031)
RECOVER SNAPSHOT did not work with replication.
Bugfix: Snapshot's UUID would override the storage one would cause the
DBMS on REPLICAS to search for the wrong database. Now in case of user
recovery, the local snapshot's UUID is overwritten.
Bugfix: Replication would not work in all cases due to the timestamp
change. Now user recovery forces a reset of the replicas.