Support regex queries with label+property indices (#3025)
Queries using a label+property index now correctly return matching
results when comparing properties against regular expressions.

The problem was caused by index code rejecting ranges across different
`PropertyType`s as being invalid. This is true in most cases, (e.g.
`WHERE node.a > 'hello' AND node.a < 42`) but not in the case where the
upper-bound is an exclusive (`<`) bound on the next (from an ordering
standpoint) `PropertyType`. As the exclusive upper bound exclusive
sentinel is not compared against, it means that the range is still
across a single `PropertyType`.

Regexes uses a range that is inclusively lower bounded on the empty
string, and exclusively upper bounded on the minimum value of whatever
`PropertyType` happens to follow string.

(This bug was a regression introduced by composite indices.)

#Fixes 3010