fix: Update grammar to allow member-access on maps in arrays
Because of the way the grammar was parsed, array subscripts would always
need to be the last part of an expression. This restriction disallowed
perfectly valid expressions like:

```cypher
RETURN [{x: 42}][0].x;
```

This fix is needed as part of GraphQL compatibility, where mutations
involving mathematical operations are represented as arrays of maps.