Allow member access on maps in lists (#3058)
Because of the way the grammar was parsed, list subscripts would always
need to be the last part of an expression. This restriction disallowed
perfectly valid expressions like:

```cypher
RETURN [{x: 42}][0].x;
```

The heirarchy of production rules in the grammar have now been changed
so that list subscript access and property member access can recursively
derive each other.

This fix is needed as part of GraphQL support
[#3027](https://github.com/memgraph/memgraph/pull/3027), where mutations
involving mathematical operations are represented as arrays of maps, and
also improves compatibility with Neo4j's Cypher.