Handle invalid accessor in the commit phase (#3065)
- Fixes invalid cached accessor in the following way: If cached accessor
is nullptr, it is impossible we should commit because replying to
prepare happens after assignment to the accessor. If cached accessor is
nullptr, and we should abort (e.g. exception was thrown while processing
deltas), we can safely return here OK because it means that the abort
already happened while destructing accessor during
ReadAndApplyDeltasSingleTxn.
- small refactoring of unique constraints in order to make debugging
easier.
- handling STRICT_SYNC replica down in Jepsen test
- removed unique constraint creation in Jepsen test since we have a bug
in GC
- a lot of comments