Support latest GraphQL middleware (#3027)
Update Memgraph to support the latest Neo4j GraphQL middleware.

- Cypher compatiblity fixes:
  - `WITH *` can now be used even if no variables are in scope.
- Subqueries no longer advance the transaction frame, fixing an issue
with vertex visibility in outer queries.
- New apoc compatible functions:
- `mgps.validate_predicate` replaces `apoc.util.validatePredicate`. This
method is used to evaluate a predicate and either error on failure or
return `true` on success.
- `date.convert_format` replaces `apoc.date.convertFormat`, and is used
to convert DateTimes to a GraphQL compatible format. (Currently, only
`iso_zoned_date_time` is supported, as this is all that GraphQL needs.)
 - Update test suite to use the latest GraphQL middleware:
   - Add extensive tests for all GraphQL mutations and queries.
- Remove the previous tests (as these were testing behaviours covered by
the new tests. No point having duplication here.)
- Remove temporary procedures used for testing. Because these query
modules are now implemented directly in Memgraph, we don't need to stub
them in test code.

---------

Co-authored-by: Dr Matt James <mattkjames7@gmail.com>
Co-authored-by: DavIvek <david.ivekovic@memgraph.io>