Add STRICT_SYNC replication mode based on two-phase commit protocol (#3026)
Jepsen HA test can now be started in sync, sync_mixed (sync+async
replicas), strict_sync and strict_sync_mixed (strict_sync + async
replicas) mode. The developer can also specify time-limit upfront.

The new type of replica is added, called `STRICT_SYNC`. The main
motivation of adding a new replica type is to support data-loss free
automatic failover. The current solution is designed to cover a variety
of failure modes and scenarios.


The cluster can be combined of either `STRICT_SYNC` and `ASYNC` or
`SYNC` and `ASYNC` replicas. The combination of `STRICT_SYNC` and `SYNC`
replicas is forbidden since such combination semantically wouldn't make
sense (`STRICT_SYNC` is a variation of `SYNC`).
The semantic the of flag `--sync-failover-only` now is better described
with: If this flag is true, you can failover to any replica except the
`ASYNC` one.

The commit method is now split into two phases called:
`PrepareForCommitPhase` and `FinalizeCommitPhase`. In the 1st phase, we
still validate existence and unique constraints same as check for
auto-indices. All deltas are made durable on disk in the 1st phase so
that instances could survive failure between two phases. A new delta is
added `WAL_TRANSACTION_START` which allows aborting durable data,
something we didn't support before. If the instance currently behaves as
`MAIN`, in the 1st phase, it will collect votes from all replicas to
determine whether the instance should be committed or not. If the
decision is yes, the value of the delta `WAL_TRANSACTION_START` will be
switched to true. The value of the delta `WAL_TRANSACTION_START` is used
in two situations:
1. When replica receives WAL file, the WAL file contains deltas both
from transactions which should be aborted and committed
2. When instance restarts, again some deltas shouldn't be applied
because they come from aborted transactions.


If all replicas voted `YES` in the 1st phase, in the 2nd phase `ldt`
(our consensus timestamp) is bumped, same as transaction's commit
timestamp. All side-effects are applied also. It is important to note
that all data is first made durable on main than on replicas, otherwise
if main fails after sending changed to replica, the replica will contain
data which main doesn't have.

The replication's state machine stayed exactly the same as it is, so all
the machinery for handling when is replica up to date or behind works as
before.

From the user's perspective, if he/she receives a message:
```
At least one STRICT_SYNC replica has not confirmed committing last transaction.
```
It means that the transaction will be aborted. This is different from
our `SYNC` mode in which we allow further committing on MAIN, although
`SYNC` replica is down.

Our RPC stack is extended with a functionality of upgrading the RPC
stream. The functionality basically allows you take only once RPC lock
and send two messages without releasing and re-taking the lock.


There are several subtleties that got changed in regards of our
durability stack that are important for future devs working on Memgraph.

1. Memgraph uses a concept of epochs to distinguish between phases in
which the different instance was MAIN. From this PR on we only push an
epoch to a history if in that epoch something was committed on the MAIN.
2. The requirement which says that there should always be a WAL file
which contains at least one delta before snapshot is taken is not true
anymore.
3. WAL file now supports interleaving of committed and aborted
transactions


Durability version is bumped to 28 but snapshots should be compatible
with older ones.

The abort of the unique constraints is moved to the common `Abort`
function.

There is only a single scenario in which users can experience a data
loss of a max one transaction which is extremely unlikely. It can happen
if the all replicas and main voted YES and after voting YES, both
replicas went down. Main commits, replies to the client and then goes
down. If replicas get back alive before main, the coordinator has no
mechanism to detect that the last txn isn't actually committed on
replicas. To support this scenario we would have to glue together
coordinators with data instances. We believe this scenario is so
unlikely that it is not worth solving at the moment.

---------

Co-authored-by: Dr Matt James <mattkjames7@gmail.com>