Fix flaky text index unit tests cleanup (#3153)
This PR removes flakiness from the text index unit test cleanup by not
failing the test if the cleanup step doesn't succeed.
A more correct approach would be to improve the handling of this type of
error in [mgcxx](https://github.com/memgraph/mgcxx), which will be added
to the TODO list for the next mgcxx release.