Speed up building by using `ccache` in CI (#3163)
The CI machines will all have a local `ccache` directory
`~/.cache/ccache` which will be mounted as a volume in the `mgbuild`
container. When the container is started, `ccache` should be installed
(using `apt` or `yum`). This should speed up building once the cache is
built up for the first time on each machine. The flag `--no-ccache` can
be used to build without it, as will be done in the coverage workflow.

An example from testing this shows that enabling `ccache` speeds up the
`Build release binary` step quite a bit from ~27 minutes to 7 minutes
33s (both Release builds on `mantis`). In this case, the cache gets
almost 100% hits because the cache is built entirely from this branch,
so more files changed will naturally mean fewer cache hits.
```
=== Ccache Statistics ===
Cacheable calls:    1025 / 1454 (70.50%)
  Hits:             1020 / 1025 (99.51%)
    Direct:         1020 / 1020 (100.0%)
    Preprocessed:      0 / 1020 ( 0.00%)
  Misses:              5 / 1025 ( 0.49%)
Uncacheable calls:   429 / 1454 (29.50%)
Local storage:
  Cache size (GiB):  0.1 /  5.0 ( 1.60%)
  Hits:             1020 / 1025 (99.51%)
  Misses:              5 / 1025 ( 0.49%)
=========================
```