Improve handling of epochs during data recovery (#3174)
When recovering from snapshot, there is no need to modify epoch history
with the epoch from snapshot since this will modify history's state
compared to the moment when snapshot is being taken.
There was a bug in one of the previous PR's in which epochs and ldt from
wal would update the state even if there was no change made in that WAL
(all txns were aborted).