Add support for exposing replication lag (#3167)
The PR adds the notion of replication lag counted based on the number of
committed txns. Main saves its own ldt and the number of committed txns
and saves the info about number of committed txns on replicas.
Coordinators can retrieve such an info through an RPC call.
We need -libatomic library for linking so we could use atomic struct.