feat: Add support for RPC versioning (#3172)
The goal of this PR is to support RPC versioning for the sake of ISSU.
RPC versioning is handled both on global level and on message level. On
the global level, we use `LoadMessageHeader` to load only appropriate
fields from the stream. The first field should always be global rpc
version after which follows the sequence of protocol-specific fields.
The current protocol specifies that after global rpc version follows
message id and message version.

Any message (Request or Response) need to specify its version. When
adding a new version for a RPC request or message, the developer should
not delete old structures but rather rename them e.g. `PrepareCommitReq`
-> `PrepareCommitReqV1` and `PrepareCommitRes` -> `PrepareCommitResV1`.
The convention is that the newest version should be called
`PrepareCommitReq` and `PrepareCommitRes`. In the new version of the
request, the developer should add a static method `Upgrade` with the
following declaration:
```
static PrepareCommitReq(PrepareCommitReqV1 const &old_req);
```

In the new version of the response, the developer should add a member
method `Downgrade` with the following declaration:
```
PrepareCommitResV1 Downgrade() const;
```

When the server receives an RPC request, it reads its version and
recursively upgrades the request until we get the request with the
newest version. The server should always reply with the response which
is of the same version as the request received. Thus, the developer
should create a response with the newest version and then recursively
downgrade it until the response with the version same as the request is
obtained.